grant the ability to request sequential test execution processing. The goal of this change to give ability to tell specs that on they own they can be executed concurrently, but when one of the dependency that is defined in a spec helper package can't be used in parallel, then request sequential processing to avoid flaky tests.
An example:
You define a business logic (use-case/interactor) specification,
and on its own, there is no side effect with the code.
But one of the dependency you need has side effect.

If you happen to have side effect free variant of that dependency for testing purpose (e.g.: in-memory isolated storage),
then you can provide that to the test and everything runs fast in concurrent execution.
This increase your local development feedback loop.

On the CI side, you still want things to be tested with real external resources,
so when the spec helper package detect that it runs in CI and it connects to an external resource,
it also can define that because this resource is used, do not allow any parallel execution.