Align behavior of Defer with Cleanup
This change is required to make integration easy between testing helpers that utilize TB#Cleanup.
Complete migration to Cleanup is not possible, because benchmark behaves differently with Cleanup,
and it would break the current behaviour of the framework.
However, it was possible to align Cleanup with After/Around and Defer teardown calls.
With this, it should not matter if Defer or Cleanup used as long it is called on `*testing.T`.