Add support for Var.OnLet
This allows the creation of generic test variable components that need to interact with the spec context.
For example a testing component that use real DB connection when an environment variable is set
can set Spec.Sequential to ensure that the test don't run on top of each other.
The usage of Var.OnLet enforce the explicit binding of Var on a Spec.