refactor fixture factory
drop the usage of context.Context and testing.TB in the Create interface.
They meant to help dynamic fixture creation across various test-cases
and thus making the Factory itself stateless.
In practice this solves an issue that might needs to be solved somewhere else,
not at the testcase framework level.