guarantee tha time generated with Random is convenient to be serialize in formats that don't support nano second precision
This is a common use-case when you work with external resources such as Postgres or MySQL.
Time is serialized majority of the time in a format that cut the nano second part.
This often cause issue when value that was persisted in a storage then retrieved
and compared with the initially generated time value to not match because the lack of nano-second precision.