adopt the latest fixture factory interface from the frameless project
The latest fixtures factory signature allows passing context awareness to the fixture creation.
If the fixture factory needs to make a type using an external resource during the test,
then the context is mandatory to properly interact with the resource.

Example:
> an entity during testing has a field that is a reference to another entity,
and without that field, the entity doesn't make sense to exists.
To avoid coupling the knowledge into these tests,
the fixture factory should be able to hide these details
when a Fixture creation is requested.