make values generated with t.Random pseudo-deterministic unique between testing scenarios
Unique values between different testing scenarios ensure that no same value is returned between different scenarios.
This behaviour is advantageous when the system under test is not accepting repeated values on consecutive interactions.
For example, a UUID is used as an entity ID in an external resource.
Using the same UUID in a consecutive would cause an error because the value was already exhausted in the previous test run.