assert.NotNil should only check if the value is not nil
Currently, it uses a try + Nil assertion to check if a value was nil,
but as a side effect, the Nil helper also prints the internals.
This could potentially result in unanticipated race conditions
for values that are perfectly fine since they aren't nil.
We should be cautious about such potential issues.